Refactored getAllUebersichte into PersonStore instead of Personenkontext
