+++ src/stores/PersonenkontextStore.spec.ts
@@ -1,6 +1,5 @@
 import {
   type DBiamPersonenkontextResponse,
-  type DBiamPersonenuebersichtControllerFindPersonenuebersichten200Response,
   type DBiamPersonenuebersichtResponse,
   type FindRollenResponse,
   type FindSchulstrukturknotenResponse,
@@ -561,62 +560,6 @@ describe('PersonenkontextStore', () => {
         personenkontextStore.getPersonenkontextAdministrationsebeneWithFilter('1', 'Org', 2);
       expect(personenkontextStore.loading).toBe(true);
       await rejects(getPersonenkontextAdministrationsebeneWithFilterPromise);
-      expect(personenkontextStore.errorCode).toEqual('some mock server error');
-      expect(personenkontextStore.loading).toBe(false);
-    });
-  });
-
-  describe('getAllPersonenuebersichte', () => {
-    it('should get All Personenuebersicht', async () => {
-      const mockResponse: DBiamPersonenuebersichtControllerFindPersonenuebersichten200Response = {
-        total: 0,
-        offset: 0,
-        limit: 0,
-        items: [
-          {
-            personId: '1234',
-            vorname: 'Samuel',
-            nachname: 'Vimes',
-            benutzername: 'string',
-            lastModifiedZuordnungen: '08.02.2024',
-            zuordnungen: [
-              {
-                sskId: 'string',
-                rolleId: 'string',
-                sskName: 'string',
-                sskDstNr: 'string',
-                rolle: 'string',
-                typ: OrganisationsTyp.Klasse,
-                administriertVon: 'string',
-                editable: true,
-                merkmale: [] as unknown as RollenMerkmal,
-              },
-            ],
-          },
-        ],
-      };
-      mockadapter.onGet('/api/dbiam/personenuebersicht').replyOnce(200, mockResponse);
-      const getAllPersonenuebersichtenPromise: Promise<void> = personenkontextStore.getAllPersonenuebersichten([]);
-      expect(personenkontextStore.loading).toBe(true);
-      await getAllPersonenuebersichtenPromise;
-      expect(personenkontextStore.allUebersichten).toEqual(mockResponse);
-      expect(personenkontextStore.loading).toBe(false);
-    });
-
-    it('should handle string error', async () => {
-      mockadapter.onGet('/api/dbiam/personenuebersicht').replyOnce(500, 'some error');
-      const getAllPersonenuebersichtenPromise: Promise<void> = personenkontextStore.getAllPersonenuebersichten([]);
-      expect(personenkontextStore.loading).toBe(true);
-      await getAllPersonenuebersichtenPromise;
-      expect(personenkontextStore.errorCode).toEqual('UNSPECIFIED_ERROR');
-      expect(personenkontextStore.loading).toBe(false);
-    });
-
-    it('should handle error code', async () => {
-      mockadapter.onGet('/api/dbiam/personenuebersicht').replyOnce(500, { code: 'some mock server error' });
-      const getAllPersonenuebersichtenPromise: Promise<void> = personenkontextStore.getAllPersonenuebersichten([]);
-      expect(personenkontextStore.loading).toBe(true);
-      await getAllPersonenuebersichtenPromise;
       expect(personenkontextStore.errorCode).toEqual('some mock server error');
       expect(personenkontextStore.loading).toBe(false);
     });
